feat: initial anime support
Not yet implemented, by far.
For now it lets you search an anime series and add it to the downloads list.

I still have to implement some specific logic, like storing both the english and romaji names of the series and episodes, as well as taking both into account when searching for episodes. Other than that, the rest should be the same as in tv shows.