feat: symlink and move
For now the app can move the episodes and movies to the correct folders and rename them. The naming schema is quite simple:
- For movies: "Movie name [Quality].extension"
- For episodes: "Series name - SxxExx [Quality].extension"

In the future I'll add the year and parse the filenames for keywords such as x364, x265, h264, h265, hevc, ac3 and such. And maybe a way to customize the pattern.

I've also added a cleaning job for torrents. There's a setting in the config that allows specifying for how long to keep a torrent after finishing the download. Torrents older that the specified amount of days will be deleted, together with the files.

Obviously, this only makes sense with symlinks.