feat: added player to the ui
First and foremost, videos in mkv format won't work in Firefox and Firefox-based browsers. They do work okay in Chromium-based browsers. Mozilla is working on adding support to mkv to their browser.

- Added a popup video player based on video.js
- The play buttons only appear in downloaded episodes and movies
- I'm working on subtitle support

This is not a substitute for Jellyfin or Plex, as it has no DLNA or native apps and won't ever be. However, it's good for watching content in a pich or just checking if the downloaded videos are correct.