feat: torznab test tool
I've added a simple tool for testing rejection filters. It accepts any indexer name and search term, and it tries to parse the season and episode number from the query.

Then it outputs random results with random terms, some of them intentionally wrong such as a few mismatched episodes or with commonly rejected terms such as HDCAM or SCREENER.

It's meant to test the rejection filters, although it can be simply used as a dummy torznab server. Refer to its readme for more info.