fix: dissappearing symlinks
The post-processor's renaming function was incorrectly overwriting existing episode symlinks by re-evaluating all files in the destination directory with the context of the latest download. This was solved by passing a specific list of newly processed files to the function, preventing it from scanning and renaming other files in the directory.

TL;DR: Old symlinks were dissappearing when postprocessing new episodes in the same destination folder.